Expose JVM specific subtypes
